requirements.txt
ctypes
win32gui
Pillow
baidu-aip
selenium
